+++ src/main/java/io/github/zeroone3010/yahueapi/Scene.java
@@ -7,6 +7,17 @@ package io.github.zeroone3010.yahueapi;
  */
 public interface Scene {
   /**
+   * <p>Returns the technical id of the scene, as assigned by the Bridge. The id stays the same even if the scene name
+   * is changed by the user.</p>
+   *
+   * <p>Note that the id is only unique within the context of a single bridge.</p>
+   *
+   * @return Id of the scene.
+   * @since 2.7.0
+   */
+  String getId();
+
+  /**
    * Activates this scene.
    */
   void activate();